[QUZ-56][FEATURE] 댓글 API 구현

+++ quiz-service/quiz-domain/src/main/kotlin/com/grepp/quizy/quiz/domain/quiz/QuizRemover.kt
@@ -1,10 +1,12 @@
 package com.grepp.quizy.quiz.domain.quiz
 
+import com.grepp.quizy.quiz.domain.useranswer.UserId
 import org.springframework.stereotype.Component
 
 @Component
 class QuizRemover(private val quizRepository: QuizRepository) {
-    fun remove(quiz: Quiz) {
+    fun remove(quiz: Quiz, deleterId: UserId) {
+        quiz.validateOwner(deleterId)
         quizRepository.delete(quiz)
     }
 }
